Refactor: move test functionality into "test" sub-library
