+++ test/web/Spec.hs
@@ -1,40 +1,18 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE NumDecimals #-}
-{-# OPTIONS_GHC -Wno-orphans #-}
 {-# LANGUAGE LambdaCase #-}
 -- WIP: unit test with "parseResult . renderResult = id"
 module Main (main) where
 
+import FunGraph.Test.Util
 import Control.Monad (forM_)
-import Control.Monad.IO.Class (liftIO)
-import Data.Data (Proxy(Proxy))
-import Data.Functor (void)
-import qualified Control.Concurrent as Conc
-import qualified Control.Concurrent as MVar
-import qualified Control.Concurrent.Async as Async
-import qualified Control.Exception as Ex
 import qualified Data.ByteString.Lazy as BSL
 import qualified Data.Set as Set
-import qualified Data.Streaming.Network
 import qualified Data.Text as T
 import qualified FunGraph
 import qualified FunGraph.Test
-import qualified Lucid
-import qualified Network.HTTP.Client
-import qualified Network.Wai
-import qualified Network.Wai.Handler.Warp
-import qualified Servant.Client
-import qualified Servant.Client.Streaming
 import qualified Server
-import qualified Server.Api
-import qualified Streaming.Prelude
 import qualified Test.Hspec as HSpec
-import Servant.API.ContentTypes
-import Servant.HTML.Lucid (HTML)
-import Server.HtmlStream (HtmlStream, toStream)
-import Test.Hspec.Expectations.Pretty (shouldBe)
 import qualified Text.HTML.Parser
 import qualified Text.HTML.Tree
 import qualified Lucid.Base
@@ -47,6 +25,7 @@ import qualified Data.Tree
 import Data.Maybe (mapMaybe)
 import qualified Data.List.NonEmpty as NE
 import qualified Types
+import qualified Server.Api
 
 testDataFileName :: FilePath
 testDataFileName = "data/all3.json"
@@ -55,12 +34,12 @@ main :: IO ()
 main =
   Server.withHandlers logger mempty testDataFileName $ \handlers ->
     runWarpTestRandomPort (Server.app handlers) $ \port -> do
-      runQuery <- mkQueryFunction port
+      queryFun <- mkQueryFunction port
+      let runQuery = fmap parsePPFunctions <$> queryFun id (Just Server.Api.NoGraph)
       main' runQuery >>= HSpec.hspec
   where
     logger = const $ pure ()
 
--- TODO: don't duplicate
 main'
   :: (FunGraph.Test.QueryTest -> IO (Either String [FunGraph.Test.PPFunctions]))
   -> IO HSpec.Spec
@@ -69,37 +48,13 @@ main' runQuery = do
         let (maxCount, _) = FunGraph.Test.queryTest_args test
         in HSpec.describe (FunGraph.Test.queryTest_name test <> " maxCount=" <> show maxCount) $ do
           HSpec.it "contained in top query results" $ do
-            eResult <- runQuery test
-            result <- either fail pure eResult
+            result <- either fail pure =<< runQuery test
             Set.fromList (map IgnorePackage result)
               `isSupersetOf`
                 Set.map IgnorePackage (FunGraph.Test.queryTest_expectedResult test)
   pure $ HSpec.describe "Integration tests" $ do
     HSpec.describe "Expected result" $
       forM_ FunGraph.Test.allTestCases testCase
-
--- TODO: don't duplicate
-isSupersetOf :: (Show a, Ord a) => Set.Set a -> Set.Set a -> IO ()
-isSupersetOf actual expected =
-  actual `shouldBe` Set.union actual expected
-
--- TODO: don't duplicate
-runWarpTestRandomPort
-  :: Network.Wai.Application -- ^ Warp 'Network.Wai.Application'
-  -> (Int -> IO ()) -- ^ Test action to run when the listening socket is ready. Argument: server port number. The server will be shut down when this IO action returns.
-  -> IO ()
-runWarpTestRandomPort app runApp = do
-  (port, socket) <- Data.Streaming.Network.bindRandomPortTCP "!4"
-  readyMVar <- MVar.newEmptyMVar
-  let warpSettings =
-        Network.Wai.Handler.Warp.setPort port $
-        Network.Wai.Handler.Warp.setBeforeMainLoop
-          (void $ Conc.forkIO $ runApp port >> Conc.putMVar readyMVar ())
-          Network.Wai.Handler.Warp.defaultSettings
-  eRes <- Async.race
-    (Network.Wai.Handler.Warp.runSettingsSocket warpSettings socket app)
-    (Conc.takeMVar readyMVar)
-  either (const $ fail "server exited") pure eRes
 
 parsePPFunctions
   :: BSL.ByteString -- ^ Entire /search endpoint response
@@ -152,72 +107,6 @@ parsePPFunctions bs = do
       _ -> False
 
     attrName (Text.HTML.Parser.Attr name _) = name
-
--- TODO: don't duplicate
-mkQueryFunction
-  :: Int
-  -> IO (FunGraph.Test.QueryTest -> IO (Either String [FunGraph.Test.PPFunctions]))
-mkQueryFunction port = do
-  manager <- Network.HTTP.Client.newManager $
-    setResponseTimeoutSeconds 60 Network.HTTP.Client.defaultManagerSettings
-  pure $ fmap parsePPFunctions . runHttpRequest manager Nothing -- TODO: also 'Just NoGraph'
-  where
-    setResponseTimeoutSeconds seconds settings =
-      settings{Network.HTTP.Client.managerResponseTimeout = Network.HTTP.Client.responseTimeoutMicro $ seconds * 1e6}
-
-    mkClientEnv manager =
-      let baseUrl' = Servant.Client.BaseUrl Servant.Client.Http "127.0.0.1" port ""
-      in Servant.Client.ClientEnv manager baseUrl' Nothing Servant.Client.defaultMakeClientRequest
-
-    runHttpRequest
-      :: Network.HTTP.Client.Manager
-      -> Maybe Server.Api.NoGraph
-      -> FunGraph.Test.QueryTest
-      -> IO BSL.ByteString
-    runHttpRequest manager mNoGraph qt =
-      let htmlStreamToStream stream = do
-            res Streaming.Prelude.:> () <- liftIO $ Streaming.Prelude.mconcat $
-              Server.HtmlStream.toStream stream
-            pure res
-          (maxCount, (src, dst)) = FunGraph.Test.queryTest_args qt
-          clientEnv = mkClientEnv manager
-          clientM = searchClientM
-            htmlStreamToStream
-            Nothing -- TODO: also bench 'HX-Boosted'?
-            (Just $ FunGraph.renderFullyQualifiedType src)
-            (Just $ FunGraph.renderFullyQualifiedType dst)
-            (Just $ fromIntegral maxCount)
-            mNoGraph
-      in Servant.Client.Streaming.runClientM clientM clientEnv >>= either Ex.throwIO pure
-
--- TODO: don't duplicate
-searchClientM
-  :: (HtmlStream IO () -> Servant.Client.Streaming.ClientM (Lucid.Html ()))
-  -> Maybe Server.Api.HxBoosted -- HX-Boosted header
-  -> Maybe T.Text -- src
-  -> Maybe T.Text -- dst
-  -> Maybe Word -- limit
-  -> Maybe Server.Api.NoGraph -- don't draw graph?
-  -> Servant.Client.Streaming.ClientM BSL.ByteString
-searchClientM consumeHtmlStream hxBoosted mSrc mDst mLimit mNoGraph =
-  Lucid.renderBS <$>
-    (queryApi hxBoosted mSrc mDst mLimit mNoGraph >>= consumeHtmlStream)
-  where
-    queryApi
-      :: Maybe Server.Api.HxBoosted
-      -> Maybe T.Text
-      -> Maybe T.Text
-      -> Maybe Word
-      -> Maybe Server.Api.NoGraph
-      -> Servant.Client.Streaming.ClientM (HtmlStream IO ())
-    queryApi =
-      Servant.Client.Streaming.client searchApi
-
-    searchApi :: Proxy Server.Api.Search
-    searchApi = Proxy
-
-instance MimeUnrender HTML (Lucid.Html ()) where
-   mimeUnrender _ = Right . Lucid.toHtmlRaw
 
 newtype IgnorePackage a = IgnorePackage a
   deriving (Show)
